Update plugin_RAMP.m
added functionality to remove peaks. Buggy but works.
Removet offset argument to find_ewave_peaks()
Got rid of warning for 'minpeakheight' in findpeaks but ended up not using that
Did a quick and dirty change to mag_click_callback() so that you can where the center of the window is.